--- src/components/Item/style.css
@@ -1,6 +1,0 @@
-.component-item {
-    display: flex;
-    border: 5px solid #000;
-    padding: 5px;
-    margin: 5px 0;
-}
