--- src/components/ItemContent/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./style.css";
-
-export default React.memo(({item}) => {
-    return (
-        <div className="component-item-content">
-            {item.content}
-        </div>
-    );
-});
