--- src/components/ItemContent/style.css
@@ -1,4 +1,0 @@
-.component-item-content {
-    flex: 1;
-    margin: 0 5px;
-}
