--- src/components/ItemDragger/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-import "./style.css";
-
-export default React.memo(() => {
-    return <div className="component-item-dragger"/>;
-});
