--- src/components/ItemDragger/style.css
@@ -1,4 +1,0 @@
-.component-item-dragger {
-    width: 50px;
-    background-color: #000;
-}
