+++ src/components/ItemList/style.css
@@ -2,3 +2,29 @@
     border: 5px solid #000;
     padding: 5px;
 }
+
+.component-item {
+    display: flex;
+    border: 5px solid #000;
+    padding: 5px;
+    margin: 5px 0;
+}
+
+.component-item-checkbox {
+    width: 10px;
+    border: 5px solid #000;
+}
+
+.component-item-checkbox[isdone=true] {
+    background-color: #00f;
+}
+
+.component-item-dragger {
+    width: 50px;
+    background-color: #000;
+}
+
+.component-item-content {
+    flex: 1;
+    margin: 0 5px;
+}
